[libc][math] Refactor exp2f implementation to header-only in src/__support/math folder.
