+++ libc/shared/math.h
@@ -48,6 +48,7 @@
 #include "math/exp10m1f.h"
 #include "math/exp10m1f16.h"
 #include "math/exp2.h"
+#include "math/exp2f.h"
 #include "math/expf.h"
 #include "math/expf16.h"
 #include "math/frexpf.h"
