+++ libc/test/shared/CMakeLists.txt
@@ -41,6 +41,7 @@ add_fp_unittest(
     libc.src.__support.math.erff
     libc.src.__support.math.exp
     libc.src.__support.math.exp2
+    libc.src.__support.math.exp2f
     libc.src.__support.math.exp10
     libc.src.__support.math.exp10f
     libc.src.__support.math.exp10f16
